Allow user to favor non-native implementation (refs #1)
